[Lect. 01] Improve RectangelCalculator

+++ src/tasks/RectangelCalculator.java
@@ -1,9 +1,9 @@
 package tasks;
 
 public class RectangelCalculator {
-    public static void calculate(int height, int width) {
-        int area = height * width;
-        int perimeter = 2 * (height + width);
+    public static void calculate(float height, float width) {
+        float area = height * width;
+        float perimeter = 2 * (height + width);
         System.out.println("Area: " + area);
         System.out.println("Perimeter: " + perimeter);
     }
